Fix Concurrency Issues With Instance Fields
In file: ConfigJsServlet.java, for a servlet class ConfigJsServlet, iCR suggested that all fields inside the class should be declared as static and/or final or simply removed. Non-static fields are not allowed. The servlet container creates one instance of each servlet for each HTTP request and the threads will share the instance fields, leading to concurrency issues.