Fix Null Dereference Check
In file: DiffUtil.java, class: DiffUtil, there is a method diffObjectDefault that, there is a potential Null pointer dereference while calling the {equals} method. iCR demonstrated the full path from the object declaration to the null dereference and fixed the issue by guarding the {equals} method call with a null check. For more information, read CWE-476.